Implemented handling of pauzes and generating 'laps' in TCX file. Should close #11 and #15
Tested on real data from Honor Band 4. When pauzing a lbs record is generated with lat=90, long=-80 and t=0
Removed the code to delete the pauze records in the validation method.
An additional list with 'lap' information at end of data dictionary (start time, stop time) when parsing the HiTrack file.
A new method to calculate the lap stats has been added. The lap stats now replace the previously global stats (start time, stop time, duration, distance)
This list with lap stats is used in the method to generate the TCX xml to filter out the records per lap.

Tested with a personal recorded walking file and imported succesfully in Strava. Help with more testing would be appreciated.

Fyi, I'm new to Python and GitHub, could be that my EDI editor or copy/paste to edit changes in GitHub changed some spacing of the original file.